seed-verifier-ml/requirements.txt
tensorflow>=2.15.0
flask==3.0.2
Pillow==10.2.0
numpy>=1.26.4